# AVX-512 accelerated batch SHA512 implementation, related features and minor cleanups