--- a/src/ballet/sha512/Local.mk
+++ b/src/ballet/sha512/Local.mk
@@ -4,6 +4,9 @@ ifdef FD_HAS_AVX
 $(call add-asms,fd_sha512_core_avx2,fd_ballet)
 $(call add-objs,fd_sha512_batch_avx,fd_ballet)
 endif
+ifdef FD_HAS_AVX512
+$(call add-objs,fd_sha512_batch_avx512,fd_ballet)
+endif
 
 $(call make-unit-test,test_sha512,test_sha512,fd_ballet fd_util)
 $(call run-unit-test,test_sha512,)