… src/ballet/sha512/fd_sha512.h
… src/ballet/sha512/fd_sha512_batch_avx512.c
… src/util/log/fd_log.c
… src/util/net/fd_pcapng.c
… src/util/scratch/fd_scratch.c
… src/util/scratch/fd_scratch.h
… src/util/scratch/test_scratch.c
… src/util/simd/fd_avx512.h
… src/util/simd/fd_avx512_wwl.h
… src/util/simd/fd_avx512_wwv.h
… src/util/simd/fd_avx_wb.h
… src/util/simd/fd_avx_wc.h
… src/util/simd/fd_avx_wi.h
… src/util/simd/fd_avx_wl.h
… src/util/simd/fd_avx_wu.h
… src/util/simd/fd_avx_wv.h
… src/util/simd/fd_sse_vb.h
… src/util/simd/fd_sse_vc.h
… src/util/simd/fd_sse_vi.h
… src/util/simd/fd_sse_vl.h
… src/util/simd/fd_sse_vu.h
… src/util/simd/fd_sse_vv.h
… src/util/simd/test_avx512_8x64.c
… src/util/tile/fd_tile.h
… src/util/tile/fd_tile_threads.cxx
… src/util/tpool/fd_tpool.cxx